pyproject.toml
[tool.poetry]
name = "fastaoc"
version = "0.1.0"
description = ""
authors = ["jonathansickert <sickert.jonathan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
ruff = "^0.8.3"
mypy = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
